Make Database type wrap all queries
Previously the database logic was limited to the history thread, so it
lived there. But now that I want to do more with it, I think I'll go
with a setup where the Database type has a mid-level Rust interface for
all of the queries it supports. With mid-level I mean that it does
accept Musium primitive types like album and track ids, but not
higher-level data structures like an index where we have to look up
string refs. That can be done at the call site. This way the Database
type can remain a relatively simple wrapper around the queries.

I am a bit anxious about passing so many arguments by position, the
probability of making a mistake is high ... but if I got it wrong, it
is will be easy to correct.

I should really add tests, but the listens aren't read from the Rust
side, so that would create even more boilerplate. For the new files
table, roundtrip tests would be good to have though.