Delete Criterion dependency and benchmark
The Criterion crate was causing a compile error, something related to
glibc that I can't be bothered to investigate right now.

I haven't used this benchmark for a long time, and I don't trust
Haskell's Criterion library, which was the direct inspiration for the
Rust version. (The reason is that they make assumptions about
distributions that often don't hold, which makes the confidence
intervals on the output misleading.) I would rather write some simple
script that dumps many timings, and analyze that with Numpy. But until
I do that, the benchmark can just be removed.

As an additional bonus, this reduces the number of dev-dependencies by a
lot.