Revert "Prevent decoding too little per burst"
This reverts commit 8d716ffd8ee7476ea65e479124f521461c1f0c0f.

I hoped that the commit would eliminate popping when starting a track,
but it did not achieve this goal, it still pops sometimes.

I do have an impression that it reduced how often this happens though,
but I did not collect any hard data on it, so I am going to assume that
a buffer underrun due to slow decode was not the cause of the popping.