Report scan status with a channel
This way we can be sure to never miss important transitions between the
different scan stages. I later want to show scan progress in the
webinterface, but we can still listen on the channel and write the
result to some "current status" variable, either atomic or protected by
a lock, from where the webinterface can serve it.