Add GitHub Actions configuration
So Travis has been dying a slow death. It was still running my CI fine,
but it turns out it ran on a Xenial machine, and one of these days it
will probably stop running my projects, so I should move away.

GitHub actions are nice, but now that I'm writing the config file for
it, I cannot help to be a little disappointed by how simple the Travis
config is in comparison to the GitHub Actions one.

Also, I probably won't get it right on the first try, but let's see if
this builds :)