Add iterator for file_metadata table full rows
This will be what should feed into the BuildMetaIndex::insert later. For
now just debug print.

Working with typed columns like this is a pain. I should really write a
macro to generate a type, and a 'Readable' and 'Bindable' impl, and a
static string with all of the columns, and one with question marks. That
would go a long way. But for now this works.