Add marcro to eliminate SQL binding boilerplate
It was a bit tiring and error-prone to first define the table with all
fields, then repeat all fields in a string literal in a query, then put
as many question marks in there as columns, then bind values, manually
specifying the order, and with the right index ... Not to mention having
to make a struct to hold them.

This should be a lot more concise (not sure if it's a net win with the
macro stuff, but it is definitely more pleasant to work with), and a lot
less sensitive to subtle typos or misorderings.

The macro is not yet used, let me get to that next.