Move index building to db::FileMetadata type
This is a step towards not indexing live, but loading from the database
instead: reading from the type that we load from the database.

In the meantime, also add a translation layer to keep the current
version working and compiling, but this will go away shortly.

I think the code can be simplified greatly, with all pattern matches and
unwraps, but that's for another commit.