Build MemoryMetaIndex from database
This is a bit hacky around progress reporting, but for now it will do
the trick. Now I can start simplifying things, and removing the
non-database path.

Performance in debug mode is surprisingly bad though ... why is it
*that* slow? Was reading 17k files from the page cache really faster? Of
course previously it was multi-threaded and now it no longer is, but I
did not expect that the workload was actually CPU-bound. (I'm not sure
if it is even now.)

But we can profile later, for now it's fast enough, a few seconds (?) in
debug mode and about a second in release mode, that's fine. Better than
a few minutes of waiting with a cold page cache.