Simplify status reporting during index building
Now that index building is relatively fast, it is not that important any
more to report problems on the fly; we can just report them all at the
end. That means we no longer need to fiddle around with threads, and
channels, we can just collect all issues in a vec.