Remove builder merging from MemoryMetaIndex
Previously I used to fill multiple builders from different threads, and
then merge them in the end. But now I no longer do that, I fill one
builder from the database, on a single thread. So there is nothing to
merge any more.

It's a bit of a shame because this code was really neat, a kind of merge
sort that also reported inconsistencies on the fly, and it was quite
readable.

But as nice as it was, this is simpler, and in most cases faster
too (multithreading did help because especially word splitting is
intensive). So it needs to go.