Make error handling a bit cleaner
This gets all of these ".expect()"s out of the way, and although it's an
improvement for the readability of the code, it might not be an
improvement at runtime, because some context that the error messages did
contain, is now lost.

None of these errors should happen during normal operation, so on the
one hand it's no big deal to lose that context. But on the other hand,
the time when you need good error messages the most is when you have
errors ...

For now I'll go with the cleaner code, and improve messages when I need
to.