Make systemd module a bit less generic
Reporting progress to systemd was not a great idea, I don't want to have
"systemctl start musium" hang for minutes (back when it would scan at
startup), and nobody cares about the status line in "systemctl status".

Signalling readiness is still useful, to know when the server is
accepting connections. But we don't need to report anything else.