Move thumbnail generation to module, handle updates
Previously, if the file existed, it would never be touched again. Now
Musium will check the mtimes, and re-generate the file if one of the
flac files in the album was newer.

Also move the machinery into a new module in the library. I want to
clean up progress reporting a bit later on.