Fix Alsa hang in playback::write_samples
I had a hang today that reproduced twice (including the initial
discovery). At the end of one particular track, playback stopped, but
the queue was not empty.

I've had hangs before but then I was unable to debug them successfully,
so since then I've been runnign Musium with debug symbols. Today I was
at least able to attach a debugger, and it turns out, the playback
thread was stuck in alsa::poll::poll, which it calls after a FillResult
::Yield. So what causes a yield? Well, one case used to be if the player
is in the prepared state, but n_consumed = 0. When n_consumed > 0, we
would start, but when n_consumed = 0, we would yield. I'm not sure why
I did that originally, I expect I must have assumed that n_consumed = 0
implies that the decoder is not yet ready, and there was nothing to
write to the buffer. But as n_consumed <= n_available, it could happen
that n_available = 0, and then we would also yield.

I don't know if this is what actually happened, but I suspect it is:
when the buffer is full, n_available = 0, and we would get stuck in a
loop and never start playback.

The solution, I think, is to not look at n_consumed, but n_available. If
n_available = 0, then the buffer is full, so we can start playback. If
n_available > 0, but our queue is empty, then we shouldn't get stuck
either, so there we can also start. And then a possibly important fall-
through: if n_available > 0 and the queue is not empty, then if
n_consumed > 0, there is something in the buffer, so we could start. But
we don't! Ask to write more instead. If we would start early, we might
underrun the buffer, so top it up completely first. I suspect this may
have had something to do with the popping sound that I sometimes hear
when starting a new track, so I am somewhat hopeful that this will also
alleviate that. To be able to diagnose that further, also add a debug
print when the PCM is in error state (which it is after an underrun).