Tweak handling of alsa::pcm::State::Prepared
If there is a tiny bit of audio in a buffer and we're not playing, we
can start and immediately drain.

If the buffer is full, there's no point returning NeedMore, we can
yield. Now that I'm thinking about it, NeedMore may even cause an
infinite loop. NeedMore is a bad name, because no more samples will ever
becaome available until we yield, the state mutex is taken, so the
decoder cannot put new samples there.