Name all threads
I noticed in gdb that not all threads were named, and I wasn't sure
where the unnamed ones are coming from. I could only find one unnamed
call to thread::spawn, and it should not happen in server mode ... but
nonetheless, it doesn't hurt to name that thread as well.