Add a state variable filter
Some modern music, especially from recent years, is super bass-heavy,
and when I play it on my HS8s, the room quickly feels “saturated”, even
at a fairly low volume. I don’t dare to turn up the volume much further,
because it is unfomfortable to listen to, but also because I don’t want
to disturb my neighbors. At this volume, the mids are too soft to enjoy
all of the details.

So I think the time has come to introduce some form of EQ. When
experimenting, statically applying a parametric EQ to remove some of
the lows around 30 Hz worked nicely. I want to be able to do this
automatically though, similar to loudness normalization. Some albums,
especially those from the 1980s, are fine as-is, but some really need
their low end toned down a bit. I’m experimenting a bit with that, but
I haven’t found a measurable metric that agrees with my subjective
experience. Either way, in the meantime I can already add this filter
here, and play with the parameters a bit.

I went down the filtering rabbit hole yesterday, from multi-tap IIR
filters, to stacked biquads, but I couldn’t get it to work nicely as a
30Hz low-pass filter with fixed point arithmetic. I later discovered the
very helpful site earlevel.com, which explains that indeed, biquadratic
filters are no good fit for low-pass filters at low cutoff frequencies,
they get numerically unstable, but the so-called “digital state variable
filter” is a good fit here. It turned out to be relatively easy to
implement, and it works great.

I prepared this in a different repository yesterday, so I could also run
some experiments, but let’s no move it here so I can embed it into the
decoding pipeline.

I also considered setting up a virtual Alsa device that has a filter
applied, but it seems pretty complex to get that working, and it would
filter everything, I would not be able to make it filter only the
bass-heavy tracks. So I think the Musium decoder is for now the best
place to put this.