Tweak filter gain to avoid clipping
Reducing by 2% is not nearly enough ... I'm not sure if there is a
theoretical upper bound that we can use here. [1] says we may need to
set the filter gain to -6 dB to avoid clipping. -6 dB is a factor 0.5
in amplitude, so it sounds like that might be something fundamental, but
I don't know why. Either way, with the previous setting I observed
strong clipping, I hope that with this I no longer do.