Bump high-pass cutoff from 40 Hz to 50 Hz
I should make this configurable, but for now, let's just rebuild.

With some of my loudest tracks, I do think with filter is better than
without, but they are still quite heavy on the bass, let's try a cutoff
of 50 Hz. This does cut a bit more of the higher bass as well, but it
also moves the entire slope, which means it cuts more from the 30 and 40
Hz ranges, than a cutoff of 40 Hz would.