Use floats for the filter after all
I wanted to do it with fixed-point arithmetic, because early generation
Raspberry Pi's don't have a hardware floating-point unit, but when
running in debug mode, I discovered even regular a 16-bit file was
causing overflow and wrapping, and I don't know of an easy way to fix
this, so let's just use floats after all. It also alleviates some of the
problems with numeric precision.