Add background scan to server, expose status
I want to be able to trigger a re-scan of the library from the
webinterface, and see the scan progress there. Add support for running a
scan in the background and read its status, and expose that in the
webinterface.

There is no way to start a scan yet, though; it needs access to the
config file. Will add that next.