Add MVar type for sharing values
The MVar (or TVar, really) is a very useful primitive in Haskell. I want
one in Rust as well. The easiest way to implement it is with a mutex,
but that is fine.

Next I can use this for the index itself in various places, and that
will ensure that the index can be updated without having to restart the
daemon.