Reload thumb cache in webserver after scan
This is the final piece of the puzzle. I can now fully rescan my library
without having to restart the server! It does still require a refresh on
the client side to make things show up though, would be nice if that
could be avoided too, but that's for future yak shaving.