Revert "Replace std locks with parking_lot locks"
This reverts commit edcf94f3c72f3a0c316e766c7d9c231bf3e94dbd.

Parking Lot has a nicer API, but there is no evidence that it improves
performance in my case, and it does increase the binary size and number
of dependencies, so I think it's best to just go with the standard
library. Some more data:

x64 binary size (release, stripped) std:         1509856 bytes
x64 binary size (release, stripped) parking_lot: 1518120 bytes

Wrk results when running against my Raspberry Pi 4 (connected to wifi,
and the client is my laptop, also connected to wifi, so wifi latency
will dominate, but perhaps there was a small effect that could emerge
from the noise).

Std:

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.88ms    4.06ms 212.10ms   96.06%
    Req/Sec   275.46     30.27   343.00     71.08%
  Latency Distribution
     50%    3.12ms
     75%    4.16ms
     90%    5.80ms
     99%   12.40ms

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.14ms    4.48ms 171.00ms   96.38%
    Req/Sec   261.04     34.39   353.00     71.84%
  Latency Distribution
     50%    3.30ms
     75%    4.50ms
     90%    6.18ms
     99%   12.93ms

Parking lot:

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.02ms    2.23ms  49.67ms   90.98%
    Req/Sec   259.41     37.82   353.00     66.76%
  Latency Distribution
     50%    3.36ms
     75%    4.49ms
     90%    6.06ms
     99%   12.55ms

  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.51ms    7.00ms 249.66ms   97.27%
    Req/Sec   252.86     37.35   343.00     75.90%
  Latency Distribution
     50%    3.37ms
     75%    4.64ms
     90%    6.57ms
     99%   18.13ms

I'm sure the parking_lot locks are better (especially because in my case
they are uncontended, and they use a spin lock for that case which
should be very fast). But the difference is insignificant in the grand
scheme of responding to the /api/queue endpoint (which takes two locks,
one to read the index MVar and one to read the player state).