Use Nixpkgs in cache key
CI is broken with a message "'cargo': No such file or directory (os
error 2". This is a misleading error, cargo does exist, it is Rustup
that fails to locate some file, I think. I think this might be because I
updated my Nixpkgs version, so now I get a newer version of Rustup, and
maybe the format is incompatible or something. Let's not reuse the
cache.