Fadvise files in loudness analysis
I have to admit, I haven't benchmarked this. Running loudness analysis
on all my files *is* slow, because of the disk, so even getting enough
of it to have sensible numbers will take me ages.

It's not a premature optimization, it is very much appropriate, because
disk access is now the bottleneck. However, I am not sure whether it is
an optimization at all. That remains to be seen ...