Always identify all missing loudness in scan
I thought I had a list of all new albums readily available, and then it
would be easy to only add those to the list, but the data is a bit more
subtle, and identifying new (or changed) albums is not obvious, so let's
just always identify everything that is missing. It's nicer anyway
because we don't need a special mode to backfill missing loudness.