Use set_buffer_size_near when switching format
The exactly 2048 worked fine so far for almost all files -- and I have
been able to play back the same file perfectly fine with it in the past
-- but now it throws an EINVAL at me. I suspect it's some change in
Alsa, related to the period size that I had to change to _near as well
recently. I haven't confirmed that this fixes the problem, but I suspect
it does.