Update to latest version of the 'alsa' crate
This brings reduced transitive dependencies ... compared to the
preceding release. But compared to the version I was on, it actually
adds a dependency :'(

This also updates the dependency on the 'nix' crate (which now becomes
purely transitive), which addresses CVE-2021-45707, although that CVE
is related to 'getgrouplist', which is not used directly in Musium, and
I don't expect it do be used indirectly through Alsa either, but I
haven't confirmed this. Let's just update it to avoid the "security
researcher" spam.