Port last remaining db interactions to Squiller
All functions that use the database are done now, but setting up the
initial schema is not. It consists of multiple statements, I will need
to add support for this to Squiller. This means that some tests fail,
and the application would be broken if you try to use it on an empty
database.