Add the possibility to store thumbs in the db
Now that I have a database anyway, and I store the waveforms and
loudness data in it, I might as well store the thumbnails there, instead
of the file system. It's one of those things that SQLite prides itself
on, being more efficient than the file system for many small files. I
don't expect it will matter much for performance (although, on spinning
disks, loading thumbnails does take a few seconds due to all the IO,
maybe it would be significantly faster with SQLite). But mostly I hope
this will simplify things by removing the need for a thumbnail
directory. Also it will allow backing up the thumbnails with the
database, and it should make it easier to properly invalidate them in
the future, because we can add more metadata than just the album id.