Tie thumbnailing together with intermediate file
This is a squash of three attempts.

I. Write thumbnails to the database (not yet fully implemented)

II. Try to use pipes for thumbnail generation

This does not work, something prints "illegal fseek", and I suspect it
is Guetzli who does need an input file and/or output file, probably it
does not like to write to pipes.

III. Use an intermediate file in a temporary directory

We still need one intermediate file between `convert` and `guetzli`,
because Guetzli cannot read from stdin, it needs seek support. But we
can put that intermediate file in $TEMPDIR, no need to read this path
from the config file. And we can write to stdout for the output, so we
can collect from there and then insert into the database. That last step
still needs to be done.