Split out tag data into separate table
I want to add support for multiple album artists per album. The current
database schema cannot represent this, as it has only one column. The
struct cannot represent it either, but first we need to fix the
database.

So let's make tags a relatively unstructured key-value affair, where
keys do not have to be unique. Then if a file has multiple ARTISTS tags,
we can store all of them.

While we're at it, we can also add a foreign key on the other tables
that can be linked back to a file. That makes it easy to invalidate this
data when we replace a file. We never do updates on the |files| table,
if a file changed, then we delete the old row and create an entirely new
one. And with the foreign keys, we can do a cascading delete and remove
all data that is potentially outdated. The only tricky one is the album
loudness, because it's derived from multiple files, but we can still
deal with that in Musium.

For now there is no migration or backwards compatibility, let's make the
new format work first, and then provide a tool to migrate.