Fix cascade delete syntax
In SQLite, unlike in Postgres, you cannot choose the cascading behavior
at delete time, you have to specify it as part of the foreign key when
creating it.

I don't like this, it makes it very implicit -- the definition of what
happens is written far away from the delete statement. But specifying it
on the delete is not supported by SQLite.