Create index on file id for tags
We need this to efficiently look up a particular tag for a given track
or file. (A linear search to then find the right tag is fine.) Also when
we are going to load the files later on, we need to find the tags by
file, so an index is justified there.