Enable searching by dual album artist
If I search for "Glasper Davis", I want to be able to find the album by
Glasper & Davis. But I don't want all their albums to show up in every
search for them. Making the artist of rank-1 in the search index when
there are mutliple album artists, is a good middle ground I think. At
least for now it seems so.