Handle new album artists schema in frontend
I can't quite say that we "handle" multiple album artists, because in a
few places we now just take the first album from the list and use that.
This needs to be extended a bit, I think the conclusion will be that the
artist pane should be able to display multiple artists at the same time.

But that's something for later.

Also delete an old unused file from the frontend source tree, the
compiler warned about an unused variable in there ... turns out the
entire file was unused!