Pack artist id array a lot tighter
Key insight: we don't need an entry per album, most albums by the same
artist can reuse the same value, as they reference a single entry ayway.

This brings down the size of the vector for my personal collection to
about the same size as the number of artists, only a few dozen more,
which is as expected.