Add note about invalidating album loudness
Although with the foreign key and cascading delete, this is not a huge
issue any more. If you update the entire album, the loudness will get
updated. Only if you somehow replace a a few tracks track in the album
(and not the one with the greatest file id), then we may fail to update
the loudness information.

I was about to write that I don’t think I ever did, but now I remember
that yesterday I merged two albums that were in the same directory but
tagged with different MusicBrainz album ids, so I guess it does happen
...