Replace import_date with first_seen
I don't care so much about the import date itself, but about what is
new. Maybe I have something in my listening history, but I modified the
file, or I did not have it in my collection for some reason and add it,
then it is not new. (Or is it?)

So mtime/import time is not the thing that matters, but when I "first
imported" a file does, which may be earlier than the mtime or import
time.