Fix stretching of full-res album art in app
I noticed popping when the full-res image loads, it was being stretched
to more than the intended height ... maybe because I added the doctype
because previously the page was triggering quirks mode (I had not heard
that term in a long time), so it was accidentally using wrong logic?

Using aspect-ratio for this instead of height seems to fix it in both
Chromium and Firefox.