Import first_seen from listens if earlier
I retagged many of my files at some point, so the mtimes on the files
are not well representative of when I really imported an album. If there
is listening data, we can get an earlier date.

At some point I should finish the ListenBrainz import, and then this can
be corrected even better, to also take into account history from before
I was using Musium.