Apply most Clippy suggestions
Or silence them, there are many where I disagree with Clippy, or where I
can see why the Clippy authors put it in, but they don't have the
context of the code and I disagree with it in how I write code. For
example, the let and returning from a block -- yes it is redundant, but
it can be very helpful to give the value a name to make the code more
self-documenting!