Make album ids 52-bit
This enables album ids to be a prefix of the track id, which means that
tracks no longer need to store a reference to their album, saving 64
bits from the Track struct. Furthermore, we already encode disc number
and track number in the track id, so we can remove those from Track as
well, saving another two bytes.

Those edits are not yet implemented in this commit though; this only
changes the format of the ids, without invasive changes elsewhere. I
also need to write the migration for this.