Drop album_id from most structs
We can now get the album id from the track id with only a single
bitshift. This shrinks the size of the (TrackId, Track) tuple down to 32
bytes, which is either a full or half a cache line on the Raspberry Pi,
and half a cache line on my 11th-gen i7. Unfortunately the alignment of
the struct itself if only 8 bytes, so there is no guarantee that the
tracks get put on a single cache line. We'll have to see if this makes a
difference and if some tweaks are needed to align the data.

In fact, we'll have to see if these changes have any impact at all,
aside from the memory savings. I'll write a benchmarking program next.