Begin working on API benchmarking script
I have a feeling that Python is going to be no good for this. (Even
measuring performance of the API might be madness ... the track lookups
may be dwarfed by all the http and json work.) Making 10 requests takes
a long time all the while my system is sitting idle. I think I need to
write my own http client and pipeline everything.