Make comparison plot more self-descriptive
Also, the results of my change to reduce the size of the Track struct
are in ... It measurably degrades performance! Not what I hoped for.
It may be that the alignment of the vector that backs the track is not
properly cache line aligned. Or maybe it's something completely
different that I missed (there is an additional shift in the binary
search now, I did not expect it to matter, but it might). I'll have to
profile to know for sure.