Shave a shift off of the get_album_tracks loop
In the migration to the new id format, I called album_id(), but that
does incur a left shift. This is unnecessary, we can just get the first
possible track id in the next album, and compare against that.

Confirmed in the disassembly, there indeed used to be a left shift
previously that is no longer there currently.

I wonder if this is what made the difference though, I can't imagine
that a single left shift could be the culprit. I also don't have my
laptop set up to do a low-noise measurement right now, so I'm just going
to commit this, look for a few other avenues for what may cause the
regression, and then do a new measurement in the end.