Identify a candidate for the regression
The bookmarks are now useless to reduce the load for scanning through
the album ids, because the top 12 bits are all zero, so all entries fall
into the zero bucket in the bookmarks!

This might very well explain the regression. I couldn't believe the
shift would do it, but messing up the album search -- I have about 1600
albums, so with functioning bookmarks, that would be 6.25 albums per
bucket, so that is three unpredictable binary search steps. But if all
1600 are in the same bucket, that is 11 unpredictable binary search
steps. If half of them are misses, it stalls the pipeline, etc., that at
least sounds more plausible than the single left shift instruction.
Also, doing a binary search through the albums table would trash the
cache which may also have an effect. The albums table is not that big,
but still 19 KB, about 5% of my CPU's L1D.