Disable alignment check for now after all
Previously it did not fire, but now it does. How weird. That means there
is potentially more performance on the table, by aligning the tracks so
they don't straddle a cache line. But then I have to put my laptop in
low-noise measurement state again -_- I don't want to do that right now.