Fix sign bug in get_album_tracks
I introduced this bug in the refactor, trying to get everything to be
efficient again after the track id refactor. This was a small change
there but it did break the get_album_tracks. This change could also
explain the benchmark result that I observed previously. Ugh, now I have
to do that all over again too.