Give tuples names so we can align them
I was a bit hesitant about this at first, it seemed like unnecessary
clutter (and it still does somewhat) in the type signatures, but on the
other hand, in the places where we iterate, I am now getting convinced
that having field names is more readable than the .0 and .1 suffix.

Performance difference remains to be measured.