Leverage bookmarks for identifying album tracks
For my current collection, this reduces the size of the slice to search
through from ~20,000 elements to 1/256th ≈ 78 elements. Instead of 14
unpredictable binary search steps, there are only 6, less than half.
This should make a big difference, possibly more than the entire TrackId
endeavor.

I'm not sure why I did not implement this earlier, all the ingredients
were there even before the id refactor and it was only a tiny change.

This change is also not benchmarked yet, to be done later when I spend
the time to get my laptop in low-noise mode again.