Begin working on playcount module
I've had this idea to count plays with an exponential moving average for
years now, and the idea to look at the difference between a moving
average with a long-ish (~years) and short-ish (~months-ish) half-life
to identify past popular tracks that I haven't listened to recently.
Let's finally start implementing this.