Write down idea for leaky bucket playcounts
I was thinking about this some more today. The current implementation
that reduces the weight if we counted recently is more or less what I
want, but also feels like a hack. And also for artists, it turns out I
don't want something that is proportional to listening time, more
something that is proportional to number of days listened.

I thought about tracking the last listen with higher resolution after
all, and what kind of formula would set the weight based on recent
listens. And then it occurred to me: just make it linear to the time
expired, a.k.a. leaky bucket rate limiting.

I have yet to implement this to see if it matches my perception of
popularity, but it feels like a really neat idea. Unlike the previous
approach, now it somehow feels natural and elegant.