Pin nanorand to version with non-broken shuffle
I was curious how the shuffle was implemented, and it turns out it was
incorrect. I thought about fixing it, but when I went to search the
repository, somebody already had. This has not been released yet, so we
pull in the library from Git. The most recent commit on main breaks Rust
1.57 compatibility, but fortunately there exists a commit in between.