Add fuzz test for shuffling
My shuffle algorithm is an excellent target for property-based testing.
I have a property: "the output is an optimal shuffle". This property is
not so easy to very though, aside from through brute force perhaps, but
we have something better. My proof (in my upcoming blog post) also gives
us the exact 2-badness that the shuffled list should have, and we can
compute this directly pretty easily from the input and from the output,
so then we can verify that the algorithm reaches the given bound.

That check is not implemented yet though.

Also include a rust-toolchain file so I remember what nightly to run
this with. I don't think rustup actually respects it in that location,
this is only for me to remember.