Implement optimality check in shuffle fuzzer
My shuffle implementation is incomplete, it doesn’t yet fix the edge
case I discovered when trying to prove it optimal. So the fuzzer should
currently be able to find counterexamples, and that's a nice way to
verify that the fuzzer works.

And behold, the fuzzer *does* find a counterexample ... but not of the
right form! This counterexample should actually be handled properly! I
added a plain test to reproduce and it does indeed reproduce. What's up?
Time to debug, but let's commit this first.