Fix bug in TestShuffler id handling
I was shifting too much. This caused the artist getter to not shift out
the album bits, so different albums were considered by defferent
artists. It was a bug in the test code, not in the shuffling algorithm.

The fuzzer still finds a counterexample, and if I minimize it, it is
also the minimal example that I came up with in my proof (and it is in
fact the smallest possible counterexample, a playlist of size 10). So
this is great, the fuzzer works and is able to find a counterexample.