Implement interleave and intersperse
I thought of a slight modification of the approach that is in the blog
post. In the blog post I pessimistically assume we have to use
intersperse, but in many cases it can be fine to use interleave, and
then we can avoid this linear scan to find adjacent bad tracks.