Rework handling of decode task for queue mutation
Now that the queue can change due to shuffle, this invariant that all
decoded data sits at the front can be violated, and furthermore, if a
decode is in progress, by the time it returns there might no longer be
the same queued track at that index.

To support this, get rid of the current_decode index. Instead we will
just loop through the queue until we find the one that's being decoded.
This linear search should not take long in practice, because there is a
memory limit on the decode buffer. Typically it's only two tracks that
have decoded blocks.

If a decode is in progress, we just ignore that, it will have to be
restarted. This now also prepares us for inserts into the queue, and
reordering.