Define table for recording ratings
I have been struggling for a while with this: I add a new album and
listen it a few times, but in the background, absentmindedly. Then one
track jumps out and I think “I really like this”. But my memory is so
bad nowadays, I cannot remember the names of those tracks later. Last.fm
has loves, and Pulsar on my phone has ratings too. I want something like
this to mark a track as "this one is nice".

So let's add a table to record this!