Pin the font for the symbol glyphs
It took some searching, but then I discovered Noto Sans Symbols 2, which
has all the glyphs I want in the right format. The heart looks a bit
wider than I would have liked, it's not the glyph that my system was
using before. But it is good enough for now. For better versions I
should use svg after all.