Replace sort arrows with larger ones
In the new font, Noto Sans Symbols 2, the glyph for the arrows that I
was using to indicate the sort direction is way too small. There is
another one that is way too big. But fortunately there is a third kind
that is about the right size. Still slightly too big for my taste, but
good enough. So switch to that one.

(Again, I should really design some icons myself and just use svgs ...)