Locate PCM device by PCM name, not card name
Suddenly (well not suddenly, after an update, but in a way where
downgrade did not fix it), opening the device started failing. It
started failing in a way where this command that I still had in my shell
history, which worked in the past, now produced an error:

    $ aplay -Dfront:0 --period-size=513 --buffer-size=2052 -c 4 -r 96000 -f S16_LE --dump-hw-params /dev/zero
    ALSA lib confmisc.c:1369:(snd_func_refer) Unable to find definition 'cards.0.pcm.front.0:CARD=0'
    ALSA lib conf.c:5181:(_snd_config_evaluate) function snd_func_refer returned error: No such file or directory
    ALSA lib conf.c:5704:(snd_config_expand) Evaluate error: No such file or directory
    ALSA lib pcm.c:2666:(snd_pcm_open_noupdate) Unknown PCM front:0
    aplay: main:834: audio open error: No such file or directory

After some debugging, the "front:0" device always fails like this now,
but we can pick a different device from "aplay --list-pcms", and that
does work, so let's do it like that in Musium as well.