Improve the discovery ranking
Based on some thoughts I had during my run, about the meaning of the exp
ratios. But then I did the math on paper, and it turns out my idea is
equivalent to an age counter that ignores playcount. So add a term
proportional to playcount after all. I have some more ideas now, I need
to work it out later. For now this is an improvement.

Also, the reverses while building the discovery rank should not have
been there. I only caught it now because the output of that ranking did
not match that of the heap-based topk one. Turns out the discoveries I
have been using since yesterday were all bogus! It does make a lot more
sense now though, these new discoveries are way better. But it does show
that maybe just random is not even a bad way to discover ...