Experiment with alternate "discover" formulas
The next thought I had: take the difference between the short-term and
long-term counter. For music played long ago, the short-term counter
will have decayed faster, so compute that factor and compensate for it
such that a track listened x time ago would come out to 0.

This does work. But then I realized that, like with my previous attempt,
the count itself does not factor into it, so this is again an age
detector that is not necessarily sensitive to absolute counts.

That said, as an age detector, this one is scarily good. With just the
difference of two exponential counters, it identifies the music that I
listened to about a year ago, but not as much in the meantime. So maybe
it does work?

Anyway, let's commit this because it is interesting.