Experiment more with discovery ranking
Taking the difference in this way does work. Scaling by sqrt(n[0]) makes
very little difference for the ranking, and the ranking is similar no
matter from which counters it is computed, so this suggests there is
something stable there, I'm not just amplifying random noise. Also when
I listen to a song they get downranked again. But still, this doesn't
feel like it is quite the right thing. Experimentation continues.